python_script/requirements.txt
requests==2.28.1
pandas==1.5.1
google-cloud-bigquery==3.3.6
